Make `OperandIter` generic over its function IL
