+++ rust/src/function.rs
@@ -397,6 +397,14 @@ impl PartialEq for Function {
     }
 }
 
+pub trait ILFunction {
+    type BNInstruction;
+    type Instruction;
+
+    fn il_instruction_from_idx(&self, expr_idx: usize) -> Self::Instruction;
+    fn operands_from_idx(&self, expr_idx: usize) -> [u64; 5];
+}
+
 /////////////////
 // AddressRange
